begin adding shopping cart issue with order id

+++ db/migrate/20170709202951_create_products.rb
@@ -3,7 +3,7 @@ class CreateProducts < ActiveRecord::Migration[5.1]
     create_table :products do |t|
       t.column :name, :string
       t.column :description, :string
-        t.column :cost, :float
+        t.column :cost, :integer
 
 
       t.timestamps
